+++ src/main/java/com/sprint/mission/discodeit/execption/ErrorCode.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.common.error;
+package com.sprint.mission.discodeit.execption;
 
 import lombok.Getter;
