+++ src/main/java/com/sprint/mission/discodeit/execption/bianryContent/BinaryContentNotNullException.java
@@ -1,7 +1,7 @@
-package com.sprint.mission.discodeit.common.error.execption.bianryContent;
+package com.sprint.mission.discodeit.execption.bianryContent;
 
-import com.sprint.mission.discodeit.common.error.ErrorCode;
-import com.sprint.mission.discodeit.common.error.execption.InvalidValueException;
+import com.sprint.mission.discodeit.execption.ErrorCode;
+import com.sprint.mission.discodeit.execption.InvalidValueException;
 
 public class BinaryContentNotNullException extends InvalidValueException {
