+++ src/main/java/com/sprint/mission/discodeit/execption/message/MessageNotFoundException.java
@@ -1,7 +1,7 @@
-package com.sprint.mission.discodeit.common.error.execption.message;
+package com.sprint.mission.discodeit.execption.message;
 
-import com.sprint.mission.discodeit.common.error.ErrorCode;
-import com.sprint.mission.discodeit.common.error.execption.EntityNotFoundException;
+import com.sprint.mission.discodeit.execption.ErrorCode;
+import com.sprint.mission.discodeit.execption.EntityNotFoundException;
 
 public class MessageNotFoundException extends EntityNotFoundException {
