+++ src/main/java/com/sprint/mission/discodeit/execption/readStatus/ReadStatusNotFoundException.java
@@ -1,7 +1,7 @@
-package com.sprint.mission.discodeit.common.error.execption.readStatus;
+package com.sprint.mission.discodeit.execption.readStatus;
 
-import com.sprint.mission.discodeit.common.error.ErrorCode;
-import com.sprint.mission.discodeit.common.error.execption.EntityNotFoundException;
+import com.sprint.mission.discodeit.execption.ErrorCode;
+import com.sprint.mission.discodeit.execption.EntityNotFoundException;
 
 public class ReadStatusNotFoundException extends EntityNotFoundException {
