+++ src/main/java/com/sprint/mission/discodeit/execption/user/UserAlreadyExistException.java
@@ -1,7 +1,7 @@
-package com.sprint.mission.discodeit.common.error.execption.user;
+package com.sprint.mission.discodeit.execption.user;
 
-import com.sprint.mission.discodeit.common.error.ErrorCode;
-import com.sprint.mission.discodeit.common.error.execption.InvalidValueException;
+import com.sprint.mission.discodeit.execption.ErrorCode;
+import com.sprint.mission.discodeit.execption.InvalidValueException;
 
 public class UserAlreadyExistException extends InvalidValueException {
