+++ src/main/java/com/sprint/mission/discodeit/execption/userStatus/UserStatusAlreadyExistException.java
@@ -1,7 +1,7 @@
-package com.sprint.mission.discodeit.common.error.execption.userStatus;
+package com.sprint.mission.discodeit.execption.userStatus;
 
-import com.sprint.mission.discodeit.common.error.ErrorCode;
-import com.sprint.mission.discodeit.common.error.execption.InvalidValueException;
+import com.sprint.mission.discodeit.execption.ErrorCode;
+import com.sprint.mission.discodeit.execption.InvalidValueException;
 
 public class UserStatusAlreadyExistException extends InvalidValueException {
