+++ src/main/java/com/sprint/mission/discodeit/validator/ChannelValidator.java
@@ -1,6 +1,6 @@
 package com.sprint.mission.discodeit.validator;
 
-import com.sprint.mission.discodeit.common.error.execption.channel.ChannelNotFoundException;
+import com.sprint.mission.discodeit.execption.channel.ChannelNotFoundException;
 import com.sprint.mission.discodeit.entity.Channel;
 import com.sprint.mission.discodeit.repository.ChannelRepository;
 import java.util.UUID;
