+++ src/main/java/com/sprint/mission/discodeit/validator/ReadStatusValidator.java
@@ -1,6 +1,6 @@
 package com.sprint.mission.discodeit.validator;
 
-import com.sprint.mission.discodeit.common.error.execption.readStatus.ReadStatusNotFoundException;
+import com.sprint.mission.discodeit.execption.readStatus.ReadStatusNotFoundException;
 import com.sprint.mission.discodeit.entity.ReadStatus;
 import com.sprint.mission.discodeit.repository.ReadStatusRepository;
 import java.util.UUID;
