+++ src/main/java/com/sprint/mission/discodeit/validator/UserStatusValidator.java
@@ -1,7 +1,7 @@
 package com.sprint.mission.discodeit.validator;
 
-import com.sprint.mission.discodeit.common.error.execption.userStatus.UserStatusAlreadyExistException;
-import com.sprint.mission.discodeit.common.error.execption.userStatus.UserStatusNotFoundException;
+import com.sprint.mission.discodeit.execption.userStatus.UserStatusAlreadyExistException;
+import com.sprint.mission.discodeit.execption.userStatus.UserStatusNotFoundException;
 import com.sprint.mission.discodeit.entity.User;
 import com.sprint.mission.discodeit.entity.UserStatus;
 import com.sprint.mission.discodeit.repository.UserStatusRepository;
